This is a Bus reservation program in C
FILENAME: PROJECT.c
DESCRIPTION: This is "Bus Reservation System" can be used in all travel
agencies.
This is very robust and user friendly application designed to reduce the
efforts of passengers. It is informative and reliable to both travel
agencies
and passengers because it stores all the informations and status about
buses.